// types/ExportData.ts
export type ExportData = {
    code: string;
    item: string;
    client: string;
    identificationNumber: string;
    city: string;
    zipCode: string;
    clientPhone: string;
    clientEmail: string;
    contactPerson: string;
    contactPersonPhone: string;
    contactPersonEmail: string;
    owner: string;
    openedFrom: string;
    status: string;
    endPrice: string;
    basicPrice: string;
    margin: string;
    category: string;
    contractManager: string;
    county: string;
    billing: string;
    source: string;
    estimatedCost: string;
    estimatedProfit: string;
}